requirements.txt
deep-translator==1.11.4
numpy==2.0.1
pandas==2.2.2
python-telegram-bot==21.4
requests==2.32.3

